web_apps/mightyish/requirements.txt
Click==7.0
Flask==1.0.3
Flask-HTTPAuth==3.3.0
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dateutil==2.8.0
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
